Move BlastActorConfiguration class out of BlastMaterial.h to its own header as they are unrelated.
