+++ Gems/Blast/Code/blast_files.cmake
@@ -8,6 +8,7 @@
 
 set(FILES
     Include/Blast/BlastActor.h
+    Include/Blast/BlastActorConfiguration.h
     Include/Blast/BlastActorData.h
     Include/Blast/BlastFamilyComponentBus.h
     Include/Blast/BlastMaterial.h
